synaptic_memristor/software/requirements.txt
pyserial
pyyaml
numpy
pandas
matplotlib
